Apply Infopen Ansible role template 0.38.0
