+++ .github/workflows/ci.yml
@@ -1,0 +1,54 @@
+---
+
+name: CI
+
+on:
+  pull_request: null
+  push:
+    branches-ignore:
+      - 'pyup-update*'
+
+jobs:
+  tests:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: [3.6, 3.7, 3.8]
+        ansible-version: [2.8, 2.9]
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v2
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v2
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -r requirements.txt
+      - name: Run tests
+        env:
+          MOLECULE_OPTIONS: ''
+          TOXENV: py${{ matrix.python-version }}-ansible${{ matrix.ansible-version }}
+        run: |
+          tox
+  release:
+    if: github.event_name == 'push' && contains(github.ref, 'refs/tags/')
+    needs:
+      - tests
+    runs-on: ubuntu-latest
+    steps:
+      - name: Upload to galaxy
+        uses: robertdebock/galaxy-action@1.0.3
+        with:
+          galaxy_api_key: ${{ secrets.galaxy_api_key }}
+      - name: Create Release
+        uses: actions/create-release@v1
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        with:
+          tag_name: ${{ github.ref }}
+          release_name: ${{ github.ref }}
+          body: ''
+          draft: false
+          prerelease: false
