+++ .pyup.yml
@@ -3,3 +3,7 @@
 # Pyup configuration
 
 branch: 'develop'
+
+requirements:
+  - 'requirements.txt'
+  - 'tox.ini'
