+++ .yamllint
@@ -1,4 +1,9 @@
+---
+
 extends: default
+
+ignore: |
+  .tox/
 
 rules:
   braces:
