+++ README.md
@@ -1,7 +1,7 @@
 # sysfs
 
-[![Build Status](https://img.shields.io/travis/infOpen/ansible-role-sysfs/master.svg?label=travis_master)](https://travis-ci.org/infOpen/ansible-role-sysfs)
-[![Build Status](https://img.shields.io/travis/infOpen/ansible-role-sysfs/develop.svg?label=travis_develop)](https://travis-ci.org/infOpen/ansible-role-sysfs)
+[![CI](https://github.com/infOpen/ansible-role-sysfs/workflows/CI/badge.svg)](https://github.com/infOpen/ansible-role-sysfs/actions)
+[![Mergify Status][mergify-status]][mergify]
 [![Updates](https://pyup.io/repos/github/infOpen/ansible-role-sysfs/shield.svg)](https://pyup.io/repos/github/infOpen/ansible-role-sysfs/)
 [![Python 3](https://pyup.io/repos/github/infOpen/ansible-role-sysfs/python-3-shield.svg)](https://pyup.io/repos/github/infOpen/ansible-role-sysfs/)
 [![Ansible Role](https://img.shields.io/ansible/role/18028.svg)](https://galaxy.ansible.com/infOpen/sysfs/)
@@ -10,25 +10,27 @@ Install sysfs package.
 
 ## Requirements
 
-This role requires Ansible 2.2 or higher,
+This role requires Ansible 2.8 or higher,
 and platform requirements are listed in the metadata file.
 
 ## Testing
 
-This role use [Molecule](https://github.com/metacloud/molecule/) to run tests.
+This role use [Molecule](https://github.com/ansible-community/molecule) to run tests.
 
-Local and Travis tests run tests on Docker by default.
+Local and Github Actions tests run tests on Docker by default.
 See molecule documentation to use other backend.
 
 Currently, tests are done on:
-- Debian Jessie
-- Ubuntu Trusty
-- Ubuntu Xenial
+- CentOS 7
+- CentOS 8
+- Debian Buster
+- Debian Stretch
+- Ubuntu Bionic
+- Ubuntu Focal
 
 and use:
-- Ansible 2.2.x
-- Ansible 2.3.x
-- Ansible 2.4.x
+- Ansible 2.8.x
+- Ansible 2.9.x
 
 ### Running tests
 
@@ -36,6 +38,14 @@ and use:
 
 ```
 $ tox
+```
+
+You can also configure molecule options and molecule command using environment variables:
+* `MOLECULE_OPTIONS` Default: "--debug"
+* `MOLECULE_COMMAND` Default: "test"
+
+```
+$ MOLECULE_OPTIONS='' MOLECULE_COMMAND=converge tox
 ```
 
 ## Role Variables
@@ -94,5 +104,8 @@ MIT
 ## Author Information
 
 Alexandre Chaussier (for Infopen company)
-- http://www.infopen.pro
+- https://www.infopen.pro
 - a.chaussier [at] infopen.pro
+
+[mergify]: https://mergify.io
+[mergify-status]: https://img.shields.io/endpoint.svg?url=https://gh.mergify.io/badges/infOpen/ansible-role-sysfs&style=flat
