+++ handlers/main.yml
@@ -7,6 +7,7 @@
   service:
     name: "{{ sysfs_service_name }}"
     state: 'restarted'
+  when: 'sysfs_service_name | bool'
   tags:
     - 'role::sysfs'
     - 'role::sysfs::install'
