+++ meta/main.yml
@@ -3,18 +3,23 @@
 galaxy_info:
   author: 'Alexandre Chaussier'
   description: 'Install and configure sysfs rules'
-  company: 'Infopen (http://www.infopen.pro)'
+  company: 'Infopen (https://www.infopen.pro)'
   license: 'MIT'
-  min_ansible_version: '2.2'
+  min_ansible_version: '2.8'
   github_branch: 'master'
   platforms:
     - name: 'Debian'
       versions:
-        - 'jessie'
+        - 'stretch'
+        - 'buster'
     - name: 'Ubuntu'
       versions:
-        - 'trusty'
-        - 'xenial'
+        - 'focal'
+        - 'bionic'
+    - name: 'EL'
+      versions:
+        - '7'
+        - '8'
   galaxy_tags:
     - 'system'
 dependencies: []
