+++ molecule/default/converge.yml
@@ -1,0 +1,8 @@
+---
+
+# Role tests
+
+- name: 'Converge | Main tasks'
+  hosts: 'all'
+  roles:
+    - role: "ansible-role-sysfs"
