--- molecule/default/playbook.yml
@@ -1,22 +1,0 @@
----
-
-# Role tests
-
-- name: 'Converge | Prerequisites tasks'
-  hosts: 'sysfs-docker-xenial'
-  gather_facts: False
-  tasks:
-    - name: 'PREREQUISITES | APT | Do an apt-get update'
-      become: True
-      raw: 'apt-get update -qq'
-      changed_when: False
-    - name: 'PREREQUISITES | APT | Install python 2.7'
-      become: True
-      raw: 'apt-get install -qq python2.7'
-      changed_when: False
-
-
-- name: 'Converge | Main tasks'
-  hosts: 'all'
-  roles:
-    - role: "ansible-role-sysfs"
