+++ tasks/manage_config.yml
@@ -3,6 +3,7 @@
 # Sys FS configuration tasks
 
 - name: 'CONFIG | Ensure rules directory exists'
+  become: True
   file:
     dest: "{{ sysfs_config.folder.path }}"
     owner: "{{ sysfs_config.folder.owner | default('root') }}"
@@ -13,6 +14,7 @@
 
 
 - name: 'CONFIG | Manage sys FS rules'
+  become: True
   template:
     src: "{{ role_path }}/templates/sysfs_rules.conf.j2"
     dest: "{{ sysfs_config.folder.path }}/{{ item.filename }}"
