+++ tasks/manage_variables.yml
@@ -21,21 +21,23 @@
 - name: 'INIT | VARIABLES | Check if OS release vars file exists'
   become: False
   stat:
-    path: "{{ role_path }}/vars/os_distribution/{{ ansible_distribution | lower }}/{{ ansible_distribution_release | lower }}.yml"
+    path: "{{ role_path }}/vars/os_distribution/{{ ansible_distribution | lower }}/{{ _os_release }}.yml"
   register: 'sysfs_check_os_release_vars'
   delegate_to: '127.0.0.1'
+  vars:
+    _os_release: "{{ (ansible_os_family | lower == 'redhat') | ternary(ansible_distribution_major_version, (ansible_distribution_release | lower)) }}"
 
 
 - name: 'INIT | VARIABLES | Load OS family vars file'
-  include_vars: "{{ role_path }}/vars/os_family/{{ ansible_os_family | lower }}.yml"
+  include_vars: "{{ sysfs_check_os_family_vars.stat.path }}"
   when: "sysfs_check_os_family_vars.stat.exists"
 
 
 - name: 'INIT | VARIABLES | Load OS distribution vars file'
-  include_vars: "{{ role_path }}/vars/os_distribution/{{ ansible_distribution | lower }}.yml"
+  include_vars: "{{ sysfs_check_os_distribution_vars.stat.path }}"
   when: "sysfs_check_os_distribution_vars.stat.exists"
 
 
 - name: 'INIT | VARIABLES | Load OS release vars file'
-  include_vars: "{{ role_path }}/vars/os_distribution/{{ ansible_distribution | lower }}/{{ ansible_distribution_release | lower }}.yml"
+  include_vars: "{{ sysfs_check_os_release_vars.stat.path }}"
   when: "sysfs_check_os_release_vars.stat.exists"
