+++ tox.ini
@@ -1,17 +1,13 @@
 [tox]
-minversion = 1.8
-envlist = py{27}-ansible{22,23,24}
+minversion = 3.16
+envlist = py{3.6,3.7,3.8}-ansible{2.8,2.9}
 skipsdist = true
 
 [testenv]
 passenv = *
 deps =
     -rrequirements.txt
-    ansible22: ansible>=2.2,<2.3
-    ansible22: docker-py==1.10.6
-    ansible23: ansible>=2.3,<2.4
-    ansible23: docker==2.5.1
-    ansible24: ansible>=2.4,<2.5
-    ansible24: docker==2.5.1
+    ansible2.8: ansible>=2.8,<2.9
+    ansible2.9: ansible>=2.9,<2.10
 commands =
-    molecule --debug test
+    molecule {env:MOLECULE_OPTIONS:--debug} {env:MOLECULE_COMMAND:test}
