+++ vars/os_family/redhat.yml
@@ -1,13 +1,13 @@
 ---
 
-# Debian family OS variables
+# RedHat family OS variables
 
 # System prerequisites
 _sysfs_packages:
   - name: 'sysfsutils'
 
 # Service management
-_sysfs_service_name: 'sysfsutils'
+_sysfs_service_name: null
 
 # Configuration management
 _sysfs_config:
